Introduce a simplest possible garbage collector
This commit implements the basis which can be used for collection of
non-referenced objects in programs written with the C programming language. The
solution is an independent implementation of a simplest possible conservative
garbage collector, fairly scanning the stack in the lookup of pointer-like
integer values which can be walked as pointers to collectable objects. The
objects in turn are also scanned if they are found by the first (stack-level)
walk. The GC well collects circular-referenced objects (in opposite to reference
counting solutions), it works in multi-threaded environments with transparent
passing of pointers between threads.

The described implementation uses one stupid lock to serialize everything
related to garbage collection and internal data structures management, so it is
very and very slow. The algorithm is also the simplest possible mark-and-sweep.

This basic implementation doesn't support scanning of .data section, thus global
pointers are not walked. It can be implemented easily but adds additional
dependency from architecture/implementation of concrete operating system. I
tried to avoid such things, even determination of stack bounds is done using
some architecture/system independent way.

There is some another problem, to support multi-threading it was necessary to
introduce a mechanism for stopping threads and performing collection from their
contexts (read "on their stacks"). It is done using user-defined signals and
some semaphore juggling. The current implementation assume that the signal uses
the same stack as the thread itself. This is not true in general case, so,
unfortunately it should be handled in future using system-dependent parsing of
signal context. Today's implementation won't work on configurations with signal
handling on separate stacks.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>